feat(scheduler): schedule with FSRS-7 instead of FSRS-6
bee-fsrs was vendored with FSRS-7 present but unwired, which made README.md's
"FSRS 7" claim half-true: the engine was in the repo, the scheduler did not call
it. LatestFsrsAdapter now reaches for Fsrs7Engine, so Kani schedules with the
35-parameter model.

Memory state needed no migration. This was the blocker that turned out not to be
one: FsrsMemoryState is shared by both engines -- the same (stability, difficulty)
pair, stability in days, difficulty on the same 1..10 scale -- so every persisted
study_items row and the state seeded from AnkiDroid during admission stayed
readable. No schema change and no state reset. Intervals do move, because the
mathematics applied to that state changed, which is the point.
FsrsVendoringTest now replays a row carrying the exact stability/difficulty the
pre-switch tests asserted for a mature card, rather than arguing the compatibility
in a comment.

Elapsed time became fractional, from one definition. KaniFsrsAdapter.review takes
a Double and FsrsElapsedTime computes it. That computation previously existed as
three hand-copied integer versions -- the review context, the adaptive engine, and
FsrsTrainingDataQueries, whose copy was annotated "exact mirror of
ReviewContext.elapsedReviewDays()". A mirror maintained by comment is one edit
from being a lie, and the failure would be near-invisible: the fitter training on
a different elapsed time than the scheduler schedules with would only surface as
a slightly wrong loss.

Scheduled intervals are floored at one day, in the adapter rather than the engine.
FSRS-6 clamped nextIntervalDays to [1, max]; FSRS-7 removes that, correctly -- an
engine should answer "when does retrievability reach the target" without assuming
what the caller does with sub-day answers. Kani's answer is that its review phase
is a day-granularity long-term queue and sub-day repetition is already
learning/relearning steps, which are practice-only. Without the floor a lapsed
card scheduled seconds out, came due in the same session, and because ladder
movement keys off the persisted FSRS due time rather than the calendar day, could
demote a rung within one session. Two existing tests caught that regression; it
is now pinned deliberately.

Personalized weights are discarded, not migrated. scheduler_fsrs_weights holds 35
values validated through Fsrs7Parameters.of, and a stored 21-value vector is
rejected like any other malformed value so the reader falls open to defaults.
There is no correct migration -- 21 weights fitted for a single power law do not
describe a blended pair -- so anyone who had opted in returns to defaults and the
weekly fitter re-earns a vector from history already on the device. This is the
one user-visible cost, and it is pinned rather than left to the generic
wrong-length case. FsrsWeightFitter carries FSRS-7's clipper bounds plus an
ordering repair for the three constraints whose bound is another parameter; a test
drives every bound past its extreme and asserts the engine accepts the result, so
the duplicated bounds table cannot drift from upstream unnoticed.

The ladder needed no change: promotesByMemoryStrength already compared
milliseconds, so the fixed-0.90 promotionIntervalDays signal and
ladder_promotion_interval_days work against fractional intervals unchanged. Those
thresholds were tuned against FSRS-6 intervals and are deliberately not retuned
here; that is a separate calibration question with its own evidence requirement.

:app needed no engine import to follow the switch -- the module boundary keeps
dev.bee.fsrs inside :core, and FsrsPersonalization re-exposes the parameter count
and defaults. The vendored tree stays byte-identical to upstream 98f3ed7.

The Fsrs7Engine.class coverage exclusion is still required, re-checked by deleting
it and rerunning :bee-fsrs:check rather than assumed: adoption made
Fsrs7Engine.create a production call, but from Kotlin, so it resolves to the
companion and leaves the @JvmStatic bridge untouched.

Four mutations confirm the new guards are not vacuous: removing the one-day floor
fails 6 tests, re-flooring the elapsed time fails 3, dropping the decay ordering
repair fails 1, and skipping weight validation fails 2.

ciFast passes uncached: 1,340 :core, 1,495 :app, and 30 :bee-fsrs tests with 0
failures, both 100% class-coverage gates green, plus 321 Python invariant tests.

+++ core/src/main/kotlin/dev/bee/kanjianki/core/KaniFsrsAdapter.kt
@@ -9,11 +9,16 @@ internal interface KaniFsrsAdapter {
         isNewLearning: Boolean,
     ): KaniFsrsReviewResult
 
+    /**
+     * [elapsedDays] is fractional; see [FsrsElapsedTime]. FSRS-7 gives a same-day
+     * review a real retrievability instead of collapsing every one onto t = 0,
+     * so flooring here would hand the engine an FSRS-6-shaped question.
+     */
     fun review(
         stability: Double,
         difficulty: Double,
         rating: String?,
-        elapsedDays: Int,
+        elapsedDays: Double,
         targetRetention: Double,
     ): KaniFsrsReviewResult
 }
