+++ core/src/test/kotlin/dev/bee/kanjianki/core/SchedulerDecisionTraceTest.kt
@@ -345,7 +345,7 @@ class SchedulerDecisionTraceTest {
             stability: Double,
             difficulty: Double,
             rating: String?,
-            elapsedDays: Int,
+            elapsedDays: Double,
             targetRetention: Double,
         ): KaniFsrsReviewResult {
             return KaniFsrsReviewResult(stability, difficulty, reviewIntervalMillis)
