+++ core/src/test/kotlin/dev/bee/kanjianki/core/SchedulerTimelineSimulatorTest.kt
@@ -444,7 +444,7 @@ class SchedulerTimelineSimulatorTest {
             stability: Double,
             difficulty: Double,
             rating: String?,
-            elapsedDays: Int,
+            elapsedDays: Double,
             targetRetention: Double,
         ): KaniFsrsReviewResult {
             return KaniFsrsReviewResult(stability, difficulty, reviewIntervalMillis)
